Replace struct 'RSAKeyPair' to namespace 'Base'
