+++ Base/Crypto/Crypto.Static/RSAKeyManager.hpp
@@ -12,7 +12,6 @@ namespace Base
 using Base::Logger::FileLogger;
 using Utility::GeneralCodes;
 using Generators::RSAKeyGenerator;
-using Models::RSAKeyPair;
 using CryptoPP::AutoSeededRandomPool;
 using CryptoPP::StringSink;
 using CryptoPP::FileSink;
